% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionannotation.R
\name{annoCivic_fusions}
\alias{annoCivic_fusions}
\title{Annotate Civic database}
\usage{
annoCivic_fusions(fusion_bedpe_dt, resource_vardir)
}
\arguments{
\item{resource_vardir}{In BioCoders resource_vardir="/home/users/patterja/BioCoders/DataResources/Variation/"
Here it is fusion_bedpe_dt (from starfusion_to_bedpe) or fusion_bedpe_annoOncokb_dt (from combineOncoKBrefs)}

\item{starfusion}{fusion_bedpe_dt data.table bedpe format}
}
\value{
fusion_bedpe_annoCivic, data.table of annotated with civic, cbind to right of bedpe
overlap is based on overlap of genomic coordinates and matching Hugo_Symbol
}
\description{
Civic table is auto uploaded 01-Mar-2017-ClinicalEvidenceSummaries.tsv
}
\examples{
fusion_bedpe_dt=starfusion_to_bedpe(starfusion)
fusion_bedpe_annoCivic_dt=annoCivic_fusions(fusion_bedpe_dt=fusion_bedpe_annoOncokb_dt, resource_vardir)
}
