% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionannotation.R
\name{annoOncoKB_fusions}
\alias{annoOncoKB_fusions}
\title{Annotate a bedpe with OncoKB combined allvars_oncokb_dt}
\usage{
annoOncoKB_fusions(allvars_oncokb_dt, fusion_bedpe_dt)
}
\arguments{
\item{allvars_oncokb_dt}{data.table combined ONCOKB variants with combineOncoKBrefs function}

\item{starfusion}{fusion_bedpe_dt data.table bedpe format with starfusion_to_bedpe function}
}
\value{
fusion_bedpe_annoOncokb, data.table annotated with oncokb, cbind to right of bedpe
overlap is based on matching Hugo_Symbol? Not so good but all we have on Oncokb
WANT TO FIX THIS ANNOTATION ONLY inverted and exact matches, no left/right only matches
}
\description{
Annotate a bedpe with OncoKB combined allvars_oncokb_dt
}
\examples{
fusion_bedpe_dt=starfusion_to_bedpe(starfusion)
actionvarsPath=paste0(resource_vardir,"OncoKB/030117/allActionableVariants.txt")
annovarsPath=paste0(resource_vardir,"OncoKB/030117/allAnnotatedVariants.txt")
allvars_oncokb_dt= combineOncoKBrefs(annovarsPath = annovarsPath, actionvarsPath = actionvarsPath)
fusion_bedpe_annoOncokb_dt=annoOncoKB_fusions(allvars_oncokb_dt = allvars_oncokb_dt, fusion_bedpe_dt = fusion_bedpe_dt)
}
