% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionannotation.R
\name{combineOncoKBrefs}
\alias{combineOncoKBrefs}
\title{combine ONCOKB references}
\usage{
combineOncoKBrefs(annovarsPath = NULL, actionvarsPath = NULL)
}
\arguments{
\item{annovarsPath}{string path to allAnnotatedVariants.txt}

\item{actionvarsPath}{string path to allActionableVariants.txt}
}
\value{
data table of combined ONCOKB variants \code{annovarsPath} and \code{actionvarsPath}
}
\description{
Combine an OncoKB annotation files (allAnnotatedVariants.txt and allActionableVariants.txt) to a combined annotation file, returns a combined data.table of ONCOKB variants.
}
\examples{
actionvarsPath=paste0(resource_vardir,"OncoKB/030117/allActionableVariants.txt")
annovarsPath=paste0(resource_vardir,"OncoKB/030117/allAnnotatedVariants.txt")
allvars_oncokb_dt= combineOncoKBrefs(annovarsPath = annovarsPath, actionvarsPath = actionvarsPath)
}
