% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionannotation.R
\name{starfusion_to_bedpe}
\alias{starfusion_to_bedpe}
\title{starfusion output to bedpe}
\usage{
starfusion_to_bedpe(starfusion)
}
\arguments{
\item{starfusion}{star-fusion.fusion_candidates.final.abridged.FFPM}
}
\value{
data.table in bedpe format
}
\description{
convert star-fusion.fusion_candidates.final.abridged.FFPM to a bedpe
}
\examples{
starfusion="/.../star-fusion.fusion_candidates.final.abridged.FFPM"
fusion_bedpe_dt=starfusion_to_bedpe(starfusion)
}
